requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai-agents>=0.0.7
tavily-python>=0.3.0
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0
eval-type-backport>=0.2.0
httpx>=0.25.0
